Reset project and create a blank slate.
The course has been updated to Angular 5
and I feel a little rusty with my skills.
I think this is a good time to work through
the course again, but this time much faster,
with better commit messages, and less of them.
I feel more confident as a developer and I
believe I will gain more from this experience
this time.